+++ packages/@unimodules/core/android/src/main/java/org/unimodules/core/interfaces/services/KeepAwakeManager.java
@@ -3,9 +3,9 @@ package org.unimodules.core.interfaces.services;
 import org.unimodules.core.errors.CurrentActivityNotFoundException;
 
 public interface KeepAwakeManager {
-  void activate(Runnable done) throws CurrentActivityNotFoundException;
+  void activate(String tag, Runnable done) throws CurrentActivityNotFoundException;
 
-  void deactivate(Runnable done) throws CurrentActivityNotFoundException;
+  void deactivate(String tag, Runnable done) throws CurrentActivityNotFoundException;
 
   boolean isActivated();
 }